+++ tests/client/unit.js
@@ -1,14 +1,14 @@
-Tinytest.add('Unit - FlowLayout._regionsToData', function(test) {
-  var data = FlowLayout._regionsToData({aa: 10, bb: "hello"});
+Tinytest.add('Unit - BlazeLayout._regionsToData', function(test) {
+  var data = BlazeLayout._regionsToData({aa: 10, bb: "hello"});
   test.equal(data.aa(), 10);
   test.equal(data.bb(), "hello");
 });
 
-Tinytest.addAsync('Unit - FlowLayout._updateRegions', function(test, done) {
+Tinytest.addAsync('Unit - BlazeLayout._updateRegions', function(test, done) {
   var aa = null;
   var bb = null;
 
-  var data = FlowLayout._regionsToData({aa: 10, bb: "hello"});
+  var data = BlazeLayout._regionsToData({aa: 10, bb: "hello"});
   var c1 = Tracker.autorun(function(c) {
     aa = data.aa();
   });
@@ -17,7 +17,7 @@ Tinytest.addAsync('Unit - FlowLayout._updateRegions', function(test, done) {
     bb = data.bb();
   });
 
-  FlowLayout._updateRegions({aa: 20});
+  BlazeLayout._updateRegions({aa: 20});
 
   Meteor.setTimeout(function() {
     test.equal(aa, 20);